+++ R/preparation.R
@@ -620,7 +620,7 @@ rankByGroup <-
 #' @examples
 #' \dontrun{
 #' # Transformation using a sliding window
-#' data.elfe2 <- rankBySlidingWindow(relfe, raw = "raw", age = "group", width = 0.5)
+#' data.elfe2 <- rankBySlidingWindow(elfe, raw = "raw", age = "group", width = 0.5)
 #'
 #' # Comparing this to the traditional approach should give us exactly the same
 #' # values, since the sample dataset only has a grouping variable for age